+++ src/server/string_family.h
@@ -63,7 +63,7 @@ class SetCmd {
 
   struct SetParams {
     uint16_t flags = SET_ALWAYS;
-    uint16_t memcache_flags = 0;
+    uint32_t memcache_flags = 0;
     uint64_t expire_after_ms = 0;     // Relative value based on now. 0 means no expiration.
     StringValue* prev_val = nullptr;  // If set, previous value is stored at pointer
